backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["nakatuba <nakatuba0626@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
Pillow = "^9.2.0"
channels = "^4.0.0"
channels-redis = "^4.0.0"
daphne = "^4.0.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"
django-model-utils = "^4.2.0"
django-storages = "^1.13.2"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
drf-nested-routers = "^0.93.4"
drf-spectacular = "^0.23.1"
factory-boy = "^3.2.1"
google-cloud-secret-manager = "^2.12.6"
google-cloud-storage = "^2.9.0"
psycopg2 = "^2.9.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.6.0"
isort = "^5.10.1"
django-stubs = "^1.12.0"
djangorestframework-stubs = "^1.7.0"

[tool.black]
skip-string-normalization = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
